[MAPSME-5027_cian_search_category] [ios] Added CIAN support to search.
