+++ iphone/Maps/UI/Search/MWMSearchManagerState.h
@@ -1,0 +1,7 @@
+typedef NS_ENUM(NSUInteger, MWMSearchManagerState)
+{
+  MWMSearchManagerStateHidden,
+  MWMSearchManagerStateDefault,
+  MWMSearchManagerStateTableSearch,
+  MWMSearchManagerStateMapSearch
+};
