+++ iphone/Maps/UI/Search/TabbedView/MWMSearchTabbedViewProtocol.h
@@ -1,6 +1,10 @@
+#import "MWMSearchManagerState.h"
+
 @protocol MWMSearchTabbedViewProtocol <NSObject>
 
 @required
+
+@property(nonatomic) MWMSearchManagerState state;
 
 - (void)searchText:(NSString *)text forInputLocale:(NSString *)locale;
 - (void)dismissKeyboard;
